Chore: moved app.module to src dir for my love @mraasvel :pikachu_dancing:
