+++ src/backend/app/src/app.module.ts
@@ -1,9 +1,9 @@
 
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AppController } from './src/app.controller';
-import { AppService } from './src/app.service';
-import { config } from "./src/orm/ormconfig";
+import { AppController } from './app.controller';
+import { AppService } from './app.service';
+import { config } from "./orm/ormconfig";
 
 @Module({
   imports: [TypeOrmModule.forRoot(config)],
